streamlit_app/requirements.txt
streamlit>=1.40.0
psycopg[binary]>=3.2.0
pandas>=2.2.0
httpx>=0.25.0
python-dotenv>=1.0.0
